% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_checks_data_to_list.R
\name{add_checks_data_to_list}
\alias{add_checks_data_to_list}
\title{Add checks dataframe to a list}
\usage{
add_checks_data_to_list(input_list_name, input_df_name)
}
\arguments{
\item{input_list_name}{List where to add the checks}

\item{input_df_name}{Checks dataframe in a global environment}
}
\value{
Updated list with added dataframe
}
\description{
Add checks dataframe to a list
}
