% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_sub_folders_setup.R
\name{add_gitignore_entries}
\alias{add_gitignore_entries}
\title{Add gitignore entries
This adds entries to the gitignore file in the project so that
the files are not followed and committed for version management}
\usage{
add_gitignore_entries(
  input_lines_to_add = c("\\n*.html", "*.xlsx", "*.xls", ".csv", ".pdf", ".doc",
    ".docx", "inputs/*", "outputs/*", "support_files/*", "!/**/.gitkeep")
)
}
\arguments{
\item{input_lines_to_add}{A vector of files and folders to ignore}
}
\description{
Add gitignore entries
This adds entries to the gitignore file in the project so that
the files are not followed and committed for version management
}
\examples{
add_gitignore_entries(input_lines_to_add = c("\n*.html", "*.xlsx", "*.xls", ".csv", ".pdf", ".doc", ".docx", "inputs/*","outputs/*","support_files/*","!/**/.gitkeep")

}
