% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_qn_label_to_cl.R
\name{add_qn_label_to_cl}
\alias{add_qn_label_to_cl}
\title{Add question label to cleaning log}
\usage{
add_qn_label_to_cl(
  input_cl,
  input_cl_name_col = "question",
  input_tool,
  input_tool_name_col = "name",
  input_tool_label_col = "label"
)
}
\arguments{
\item{input_cl}{Specify the data frame for the cleaning log}

\item{input_cl_name_col}{Specify the question/name column of cleaning log}

\item{input_tool}{Specify the data frame for the survey tool}

\item{input_tool_name_col}{Specify the name column of the tool}

\item{input_tool_label_col}{Specify the label column of the tool}
}
\description{
Add question label to cleaning log
}
