% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_select_rename.R
\name{batch_select_rename}
\alias{batch_select_rename}
\title{Batch select and rename columns}
\usage{
batch_select_rename(
  input_df,
  input_selection_str = "i.check.",
  input_replacement_str = ""
)
}
\arguments{
\item{input_replacement_str}{}
}
\description{
Batch select and rename columns
}
