% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_spatial.R
\name{check_duplicate_pt_numbers}
\alias{check_duplicate_pt_numbers}
\title{Check duplicate point numbers}
\usage{
check_duplicate_pt_numbers(input_tool_data, input_sample_pt_nos_list)
}
\arguments{
\item{input_tool_data}{Specify the data frame for the tool data}

\item{input_sample_pt_nos_list}{Specify a list of sample point numbers}
}
\value{
Data frame of surveys with duplicate point_number
}
\description{
Check duplicate point numbers
}
