% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_outliers.R
\name{check_outliers_cleaninginspector_repeats}
\alias{check_outliers_cleaninginspector_repeats}
\title{Check for outliers with repeats using cleaninginspectoR package and format the log}
\usage{
check_outliers_cleaninginspector_repeats(
  input_tool_data,
  input_sheet_name,
  input_repeat_cols
)
}
\arguments{
\item{input_tool_data}{Specify the data frame for the repeat loop data joined with main dataset}

\item{input_sheet_name}{Specify the sheet name as defined in the tool}

\item{input_repeat_cols}{Specify the column names in the repeat loop to be checked}
}
\value{
The resulting data frame of data out of range
}
\description{
Check for outliers with repeats using cleaninginspectoR package and format the log
}
