% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_outliers.R
\name{check_outliers_on_column}
\alias{check_outliers_on_column}
\title{Check for outliers on a column using percentiles}
\usage{
check_outliers_on_column(
  input_tool_data,
  input_column,
  input_lower_limit,
  input_upper_limit
)
}
\arguments{
\item{input_tool_data}{Specify the data frame for the tool data}

\item{input_column}{Specify the column where to detect outliers}

\item{input_lower_limit}{Lower limit value of the lower percentile}

\item{input_upper_limit}{Upper limit value of the upper percentile}
}
\value{
The resulting data frame of data out of range
}
\description{
Check for outliers on a column using percentiles
}
