% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_outliers.R
\name{check_outliers_on_column_repeats}
\alias{check_outliers_on_column_repeats}
\title{Check for outliers on a column in a repeat loop using percentiles}
\usage{
check_outliers_on_column_repeats(
  input_tool_data,
  input_enumerator_id_col = "enumerator_id",
  input_location_col,
  input_column,
  input_lower_limit,
  input_upper_limit,
  input_sheet_name
)
}
\arguments{
\item{input_tool_data}{Specify the data frame for the repeat loop data joined with main dataset}

\item{input_enumerator_id_col}{Specify the enumerator id column}

\item{input_location_col}{Specify the location description column}

\item{input_column}{Specify the column where to detect outliers}

\item{input_lower_limit}{Lower limit value of the lower percentile}

\item{input_upper_limit}{Upper limit value of the upper percentile}

\item{input_sheet_name}{Specify the sheet name as defined in the tool}
}
\value{
The resulting data frame of data out of range
}
\description{
Check for outliers on a column in a repeat loop using percentiles
}
