% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_spatial.R
\name{check_pt_number_not_in_samples}
\alias{check_pt_number_not_in_samples}
\title{Check for point numbers not in samples}
\usage{
check_pt_number_not_in_samples(input_tool_data, input_sample_pt_nos_list)
}
\arguments{
\item{input_tool_data}{Specify the data frame for the tool data}

\item{input_sample_pt_nos_list}{Specify a list of sample point numbers}
}
\value{
Data frame of surveys with specified point numbers that are not in the samples
}
\description{
Check for point numbers not in samples
}
