% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_survey_time.R
\name{check_survey_time}
\alias{check_survey_time}
\title{Check survey time against expected minimum time and maximum time of the survey}
\usage{
check_survey_time(input_tool_data, input_min_time, input_max_time)
}
\arguments{
\item{input_tool_data}{Specify the data frame for the tool data}

\item{input_min_time}{Specify the minimum expected time for the survey}

\item{input_max_time}{Specify the maximum expected time for the survey}
}
\value{
The resulting data frame of surveys outside of the minimum and maximum times
}
\description{
Check survey time against expected minimum time and maximum time of the survey
}
