% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_spatial.R
\name{check_threshold_distance}
\alias{check_threshold_distance}
\title{check that collected point is not at a distance greater than the threshold}
\usage{
check_threshold_distance(
  input_sample_data,
  input_tool_data,
  input_enumerator_id_col = "enumerator_id",
  input_location_col,
  input_point_id_col,
  input_threshold_dist = 150,
  input_geopoint_col = "geopoint"
)
}
\arguments{
\item{input_sample_data}{Specify a GIS layer with sample data}

\item{input_tool_data}{Specify the data frame for the tool data with "_geopoint_longitude" and "_geopoint_latitude" columns}

\item{input_enumerator_id_col}{Specify the enumerator id column}

\item{input_location_col}{Specify the location description column}

\item{input_point_id_col}{Specify the point id column}

\item{input_threshold_dist}{Specify threshold distance. Default is 150m}

\item{input_geopoint_col}{Specify column name for the geopoint}
}
\value{
Data frame with surveys collected at a distance greater than the specified threshold distance
}
\description{
check that collected point is not at a distance greater than the threshold
}
