% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_survey_time.R
\name{check_time_interval_btn_surveys}
\alias{check_time_interval_btn_surveys}
\title{Check interval between surveys by the same enumerator}
\usage{
check_time_interval_btn_surveys(
  input_tool_data,
  input_enumerator_id_col = "enumerator_id",
  input_location_col,
  input_min_time
)
}
\arguments{
\item{input_tool_data}{Specify the data frame for the tool data}

\item{input_enumerator_id_col}{Specify the enumerator id column}

\item{input_location_col}{Specify the location description column}

\item{input_min_time}{Specify the minimum expected time for an enumerator to move from one survey location to another}
}
\value{
The resulting data frame of surveys not meeting this requirement
}
\description{
Check interval between surveys by the same enumerator
}
