% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_add_extra_cols.R
\name{checks_add_extra_cols}
\alias{checks_add_extra_cols}
\title{Add extra columns to harmonise checking
This adds columns uuid, start_date, enum_id_col and location_col to the data}
\usage{
checks_add_extra_cols(
  input_tool_data,
  input_enumerator_id_col,
  input_location_col
)
}
\arguments{
\item{input_tool_data}{Specify the data frame for the tool data}

\item{input_enumerator_id_col}{Specify the enumerator id column}

\item{input_location_col}{Specify the location description column}
}
\description{
Add extra columns to harmonise checking
This adds columns uuid, start_date, enum_id_col and location_col to the data
}
