% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_support.R
\name{cleaning_support}
\alias{cleaning_support}
\title{Cleaning support
Uses packages like kobold and butteR}
\usage{
cleaning_support(
  input_df_raw_data,
  input_df_survey,
  input_df_choices,
  input_df_cleaning_log,
  input_vars_to_remove_from_data = c("deviceid", "audit", "audit_URL", "instance_name",
    "complainant_name", "complainant_id", "respondent_telephone", "name_pers_recording"),
  input_geopoint_col = "geopoint"
)
}
\arguments{
\item{input_df_raw_data}{Specify the data frame for the tool data}

\item{input_df_survey}{Specify the data frame for the survey sheet}

\item{input_df_choices}{Specify the data frame for the choices sheet}

\item{input_df_cleaning_log}{Specify the data frame for the cleaning/checking log}

\item{input_vars_to_remove_from_data}{Specify a vector of columns to blank. This is especially to handle PII columns in the dataset.}

\item{input_geopoint_col}{Specify column name for the geopoint}
}
\value{
Data frame of cleaned data
}
\description{
Cleaning support
Uses packages like kobold and butteR
}
