% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_sub_folders_setup.R
\name{create_dot_gitkeep}
\alias{create_dot_gitkeep}
\title{Add .gitkeep file to ignored sub folders in .gitignore}
\usage{
create_dot_gitkeep(input_sub_folders = c("inputs", "outputs"))
}
\arguments{
\item{input_sub_folders}{Sub folders to be maintained as empty folders in the project structure while committing
The default folders are "inputs", "outputs"}
}
\description{
Add .gitkeep file to ignored sub folders in .gitignore
}
\examples{
create_dot_gitkeep(input_sub_folders = c("inputs", "outputs"))
# This will add .gitignore in the sub folders "inputs", "outputs"

}
