% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project_sub_folders.R
\name{create_project_sub_folders}
\alias{create_project_sub_folders}
\title{Create sub folders for a project in RStudio}
\usage{
create_project_sub_folders(
  input_sub_folders = c("inputs", "outputs", "R", "support_files")
)
}
\arguments{
\item{input_sub_folders}{a vector of strings representing names of folders.
The default folders are c("inputs", "outputs", "R", "support_files")}
}
\description{
Create sub folders for a project in RStudio
}
\examples{
create_project_sub_folders(input_sub_folders = c("inputs", "outputs", "R", "support_files"))
This will create the sub folders folders "inputs", "outputs", "R", "support_files" in the current R project

}
