% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_support.R
\name{cts_add_new_sm_choices_to_data}
\alias{cts_add_new_sm_choices_to_data}
\title{Add new select multiple choices to the data
This function adds new choices to be added on sm questions after data collection}
\usage{
cts_add_new_sm_choices_to_data(
  input_df_tool_data,
  input_df_filled_cl,
  input_df_survey,
  input_df_choices,
  input_sm_seperator = "/"
)
}
\arguments{
\item{input_df_tool_data}{Specify the data frame for the raw data}

\item{input_df_filled_cl}{Specify the data frame for the filled cleaning log}

\item{input_df_survey}{Specify the data frame for the survey sheet}

\item{input_df_choices}{Specify the data frame for the choices sheet}

\item{input_sm_seperator}{The seperator for select multiple}
}
\value{
An updated data frame of the data with added columns for new choices
}
\description{
Add new select multiple choices to the data
This function adds new choices to be added on sm questions after data collection
}
