% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_spatial.R
\name{cts_check_pt_number_not_in_samples}
\alias{cts_check_pt_number_not_in_samples}
\title{Cleaningtools format check for point numbers not in samples}
\usage{
cts_check_pt_number_not_in_samples(
  input_tool_data,
  input_uuid_col = "_uuid",
  input_point_id_col,
  input_sample_pt_nos_list
)
}
\arguments{
\item{input_tool_data}{Specify the data frame for the tool data}

\item{input_uuid_col}{Specify the uuid column}

\item{input_point_id_col}{Specify the point id column}

\item{input_sample_pt_nos_list}{Specify a list of sample point numbers}
}
\value{
Data frame of surveys with specified point numbers that are not in the samples
}
\description{
Cleaningtools format check for point numbers not in samples
}
