% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_other_specify_data.R
\name{cts_format_other_specify_repeats}
\alias{cts_format_other_specify_repeats}
\title{Cleaningtools format other specify repeats}
\usage{
cts_format_other_specify_repeats(
  input_repeat_data,
  input_uuid_col = "_submission__uuid",
  input_survey,
  input_choices,
  input_sheet_name,
  input_index_col = "_index"
)
}
\arguments{
\item{input_repeat_data}{Specify the data frame for the repeat data.}

\item{input_uuid_col}{Specify the uuid column}

\item{input_survey}{Specify the data frame for the survey sheet}

\item{input_choices}{Specify the data frame for the choices sheet}

\item{input_sheet_name}{Specify the sheet name as in the tool}

\item{input_index_col}{Specify the index column in the repeat data}
}
\description{
Cleaningtools format other specify repeats
}
