% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_other_specify_data.R
\name{cts_other_specify}
\alias{cts_other_specify}
\title{Cleaningtools format other specify}
\usage{
cts_other_specify(
  input_tool_data,
  input_uuid_col = "_uuid",
  input_survey,
  input_choices
)
}
\arguments{
\item{input_tool_data}{Specify the data frame for the tool data}

\item{input_uuid_col}{Specify the uuid column}

\item{input_survey}{Specify the data frame for the survey sheet}

\item{input_choices}{Specify the data frame for the choices sheet}
}
\description{
Cleaningtools format other specify
}
\examples{
cts_other_specify(input_tool_data = df_tool_data,
                         input_uuid_col = "_uuid",
                         input_survey = df_survey,
                         input_choices = df_choices)

}
