% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_support.R
\name{cts_update_sm_parent_cols}
\alias{cts_update_sm_parent_cols}
\title{Update select multiple parent columns
This function updates select multiple parent columns trying to keep the original selection order}
\usage{
cts_update_sm_parent_cols(
  input_df_cleaning_step_data,
  input_sm_seperator = "/",
  input_uuid_col = "uuid",
  input_enumerator_id_col = "enumerator_id",
  input_point_id_col,
  input_collected_date_col,
  input_location_col,
  input_dataset_type = "main",
  input_sheet_name,
  input_index_col
)
}
\arguments{
\item{input_df_cleaning_step_data}{The output data frame from the cleaning step}

\item{input_sm_seperator}{The seperator for select multiple}

\item{input_uuid_col}{Specify the uuid column}

\item{input_enumerator_id_col}{Specify the enumerator id column}

\item{input_point_id_col}{Specify the point id column}

\item{input_collected_date_col}{Specify the column for the date of data collection}

\item{input_location_col}{Specify the location description column}

\item{input_dataset_type}{Specify the dataset type. Options are 'main' or 'loop'}

\item{input_sheet_name}{Specify the sheet name as in the tool}

\item{input_index_col}{Specify the index column in the repeat data}
}
\value{
A list that contains an updated data and the extra log to add to the original log used for cleaning
}
\description{
Update select multiple parent columns
This function updates select multiple parent columns trying to keep the original selection order
}
