% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_other_specify_data.R
\name{extract_other_specify_data}
\alias{extract_other_specify_data}
\title{Title}
\usage{
extract_other_specify_data(input_tool_data, input_survey, input_choices)
}
\arguments{
\item{input_tool_data}{Specify the dataframe for the tool data}

\item{input_survey}{Specify the dataframe for the survey sheet}

\item{input_choices}{Specify the dataframe for the choices sheet}
}
\description{
Title
}
