% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_survey_time.R
\name{get_average_survey_time}
\alias{get_average_survey_time}
\title{Get average survey time
This function tends to exclude outliers}
\usage{
get_average_survey_time(
  input_tool_data,
  input_lower_limit = 0.025,
  input_upper_limit = 0.975
)
}
\arguments{
\item{input_tool_data}{Specify the data frame for the tool data}

\item{input_lower_limit}{Specify the quantile lower limit}

\item{input_upper_limit}{Specify the quantile upper limit}
}
\description{
Get average survey time
This function tends to exclude outliers
}
