% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_maps.R
\name{map_with_ggplot}
\alias{map_with_ggplot}
\title{map_with_ggplot Create a map with ggplot}
\usage{
map_with_ggplot(input_shepefile, input_col, input_ctry_code)
}
\arguments{
\item{input_shepefile}{The shapefile/layer to visulise}

\item{input_col}{Column for styling the map. This column should be present in the layer provided}

\item{input_ctry_code}{Provide country code/location of the map}
}
\description{
map_with_ggplot Create a map with ggplot
}
