% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_maps.R
\name{map_with_tmap}
\alias{map_with_tmap}
\title{map_with_tmap
Function to iterate over given columns and produce maps based on those columns}
\usage{
map_with_tmap(input_shepefile, input_col, input_ctry_code)
}
\arguments{
\item{input_shepefile}{The shapefile/layer to visulise}

\item{input_col}{Column for styling the map. This column should be present in the layer provided}

\item{input_ctry_code}{Provide country code/location of the map}
}
\description{
map_with_tmap
Function to iterate over given columns and produce maps based on those columns
}
