% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode_with_nc.R
\name{mode_with_nc}
\alias{mode_with_nc}
\title{Calculate mode with No Consensus for ties}
\usage{
mode_with_nc(x, na.rm = FALSE)
}
\arguments{
\item{x}{Specify a vector whose mode is to be calculated}

\item{na.rm}{}
}
\description{
Calculate mode with No Consensus for ties
}
