% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_points_for_kobo.R
\name{samples_for_kobo}
\alias{samples_for_kobo}
\title{Sample points for kobo
This function prepares coordinates and other columns for adding external data
to kobo to support spatial checks during data collection}
\usage{
samples_for_kobo(
  input_gis_pt_layer,
  input_pt_no_col = "pt_num",
  output_longitude_col = "longitude",
  output_latitude_col = "latitude",
  output_description_col = "description",
  output_file_name = "sample_pts_data",
  output_folder_name = "outputs"
)
}
\arguments{
\item{input_gis_pt_layer}{Specify the layer where to extract data}

\item{input_pt_no_col}{Specify the input point number column}

\item{output_longitude_col}{Specify the longitude column to output}

\item{output_latitude_col}{Specify the latitude column to output}

\item{output_description_col}{Specify the description column to output}

\item{output_file_name}{Specify the file name to output}

\item{output_folder_name}{Specify the folder where to export data}
}
\description{
Sample points for kobo
This function prepares coordinates and other columns for adding external data
to kobo to support spatial checks during data collection
}
