% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_sub_folders_setup.R
\name{add_dot_gitkeep_to_folder}
\alias{add_dot_gitkeep_to_folder}
\title{Add .gitkeep file to ignored sub folders in .gitignore}
\usage{
add_dot_gitkeep_to_folder(input_sub_folders = c("inputs", "outputs"))
}
\arguments{
\item{input_sub_folders}{Sub folders to be maintained as empty folders in the project structure while committing
The default folders are "inputs", "outputs"}
}
\description{
Add .gitkeep file to ignored sub folders in .gitignore
}
\examples{
add_dot_gitkeep_to_folder(input_sub_folders = c("inputs", "outputs"))
# This will add .gitkeep in the sub folders "inputs", "outputs"

}
