% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_duplicate_uuids.R
\name{cts_checks_duplicate_uuids}
\alias{cts_checks_duplicate_uuids}
\title{Cleaningtools format check data for duplicate uuids}
\usage{
cts_checks_duplicate_uuids(input_tool_data, input_uuid_col = "_uuid")
}
\arguments{
\item{input_tool_data}{Specify the data frame for the tool data}

\item{input_uuid_col}{Specify the uuid column}
}
\description{
Cleaningtools format check data for duplicate uuids
}
