% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_other_specify_data.R
\name{cts_format_other_specify}
\alias{cts_format_other_specify}
\title{Cleaningtools format other specify}
\usage{
cts_format_other_specify(
  input_tool_data,
  input_enumerator_id_col = "enumerator_id",
  input_location_col = "district_name",
  input_survey,
  input_choices
)
}
\arguments{
\item{input_tool_data}{Specify the data frame for the tool data}

\item{input_enumerator_id_col}{Specify the enumerator id column}

\item{input_location_col}{Specify the column for location description in the dataset}

\item{input_survey}{Specify the data frame for the survey sheet}

\item{input_choices}{Specify the data frame for the choices sheet}
}
\description{
Cleaningtools format other specify
}
\examples{
cts_format_other_specify(input_repeat_data = df_tool_data_protection_risky_places,
                         input_enumerator_id_col = "enumerator_id",
                         input_location_col = "district_name",
                         input_survey = df_survey,
                         input_choices = df_choices)

}
