% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_support.R
\name{cts_update_sm_parent_cols}
\alias{cts_update_sm_parent_cols}
\title{Update select multiple parent columns
This function updates select multiple parent columns trying to keep the original selection order}
\usage{
cts_update_sm_parent_cols(
  input_df_cleaning_step_data,
  input_sm_seperator = "/",
  input_location_col
)
}
\arguments{
\item{input_df_cleaning_step_data}{The output data frame from the cleaning step}

\item{input_sm_seperator}{The seperator for select multiple}

\item{input_location_col}{Specify the location description column}
}
\value{
A list that contains an updated data and the extra log to add to the original log used for cleaning
}
\description{
Update select multiple parent columns
This function updates select multiple parent columns trying to keep the original selection order
}
