% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_other_specify_data.R
\name{extract_other_specify_data_repeats}
\alias{extract_other_specify_data_repeats}
\title{Extract other specify data for repeats}
\usage{
extract_other_specify_data_repeats(
  input_repeat_data,
  input_survey,
  input_choices,
  input_sheet_name,
  input_repeat_cols
)
}
\arguments{
\item{input_repeat_data}{Specify the data frame for the main dataset joined with repeat. Keep the order Main dataset then join repeat}

\item{input_survey}{Specify the data frame for the survey sheet}

\item{input_choices}{Specify the data frame for the choices sheet}

\item{input_sheet_name}{Specify the sheet name as in the tool}

\item{input_repeat_cols}{Specify the columns in the repeat that have other specify}
}
\value{
A data frame of other specify for the repeats
}
\description{
Extract other specify data for repeats
}
\examples{
extract_other_specify_data_repeats(input_repeat_data = df_tool_data_protection_risky_places,
                                   input_survey = df_survey,
                                   input_choices = df_choices,
                                   input_sheet_name = "protection_risky_places",
                                   input_repeat_cols = c("places_where_children_are_mostly_at_risk"))

}
